chore: simplify
Run documentLinks only on goto_file invocation.
