+++ helix-term/src/handlers.rs
@@ -13,13 +13,11 @@ use crate::handlers::signature_help::SignatureHelpHandler;
 pub use helix_view::handlers::{word_index, Handlers};
 
 use self::document_colors::DocumentColorsHandler;
-use self::document_links::DocumentLinksHandler;
 
 mod auto_save;
 pub mod completion;
 pub mod diagnostics;
 mod document_colors;
-mod document_links;
 mod prompt;
 mod signature_help;
 mod snippet;
@@ -31,7 +29,6 @@ pub fn setup(config: Arc<ArcSwap<Config>>) -> Handlers {
     let signature_hints = SignatureHelpHandler::new().spawn();
     let auto_save = AutoSaveHandler::new().spawn();
     let document_colors = DocumentColorsHandler::default().spawn();
-    let document_links = DocumentLinksHandler::default().spawn();
     let word_index = word_index::Handler::spawn();
     let pull_diagnostics = PullDiagnosticsHandler::default().spawn();
     let pull_all_documents_diagnostics = PullAllDocumentsDiagnosticHandler::default().spawn();
@@ -41,7 +38,6 @@ pub fn setup(config: Arc<ArcSwap<Config>>) -> Handlers {
         signature_hints,
         auto_save,
         document_colors,
-        document_links,
         word_index,
         pull_diagnostics,
         pull_all_documents_diagnostics,
@@ -54,7 +50,6 @@ pub fn setup(config: Arc<ArcSwap<Config>>) -> Handlers {
     diagnostics::register_hooks(&handlers);
     snippet::register_hooks(&handlers);
     document_colors::register_hooks(&handlers);
-    document_links::register_hooks(&handlers);
     prompt::register_hooks(&handlers);
     handlers
 }
