+++ helix-view/src/document.rs
@@ -208,13 +208,10 @@ pub struct Document {
 
     /// Annotations for LSP document color swatches
     pub color_swatches: Option<DocumentColorSwatches>,
-    /// Cached LSP document links for navigation (e.g. goto_file).
-    pub document_links: Vec<DocumentLink>,
     // NOTE: ideally this would live on the handler for color swatches. This is blocked on a
     // large refactor that would make `&mut Editor` available on the `DocumentDidChange` event.
     pub color_swatch_controller: TaskController,
     pub pull_diagnostic_controller: TaskController,
-    pub document_link_controller: TaskController,
 
     // NOTE: this field should eventually go away - we should use the Editor's syn_loader instead
     // of storing a copy on every doc. Then we can remove the surrounding `Arc` and use the
@@ -741,12 +738,10 @@ impl Document {
             readonly: false,
             jump_labels: HashMap::new(),
             color_swatches: None,
-            document_links: Vec::new(),
             color_swatch_controller: TaskController::new(),
             syn_loader,
             previous_diagnostic_id: None,
             pull_diagnostic_controller: TaskController::new(),
-            document_link_controller: TaskController::new(),
         }
     }
