+++ helix-view/src/handlers/lsp.rs
@@ -16,8 +16,6 @@ use helix_lsp::{lsp, LanguageServerId, OffsetEncoding};
 use super::Handlers;
 
 pub struct DocumentColorsEvent(pub DocumentId);
-pub struct DocumentLinksEvent(pub DocumentId);
-
 #[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum SignatureHelpInvoked {
     Automatic,
